Move components to separate file
